+++ source/sw_main.c
@@ -1361,7 +1361,7 @@ void	Draw_FillEx( int x, int y, int w, int h, const color_t color );
 void	Draw_StretchRaw( int x, int y, int w, int h, int cols,
             int rows, const byte *data );
 void    Draw_Char( int x, int y, int flags, int ch, qhandle_t hFont );
-int 	Draw_String( int x, int y, int flags, int maxChars,
+int 	Draw_String( int x, int y, int flags, size_t maxChars,
             const char *string, qhandle_t hFont );
 
 /*
